% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\docType{data}
\name{Builder}
\alias{Builder}
\title{Build a Graphic}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Builder
}
\description{
Builds and renders a \link{Graphic} using a \link{Renderer}.
}
\keyword{datasets}
