% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascading-theme.R
\name{cascading_element}
\alias{cascading_element}
\title{Create a cascading element}
\usage{
cascading_element(..., .subclass = NULL)
}
\arguments{
\item{...}{Key/value pairs. Passed to \code{\link[rlang:list2]{rlang::list2()}}, so
tidy evaluation is supported.}

\item{.subclass}{The class to use for this cascading element}
}
\value{
A list with class \code{cascading_element} and any subclasses
specified by \code{.subclass}.
}
\description{
A cascading element is a \code{\link[=list]{list()}} with unique names whose elements
can be inherited by a child \code{\link[=cascading_element]{cascading_element()}}. The details of
cascading are handled by the \code{\link[=cascade]{cascade()}} and \code{\link[=reverse_cascade]{reverse_cascade()}}
generics.
}
