% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascading-theme.R
\name{CascadingTheme}
\alias{CascadingTheme}
\title{Cascading Themes}
\description{
Strong graphics packages are built on good defaults that are
highly customizable. The \link{CascadingTheme} class makes it
possible to define a list of defaults that specify inheritance,
allowing customizable defaults with less repeated code.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CascadingTheme$new()}}
\item \href{#method-compile}{\code{CascadingTheme$compile()}}
\item \href{#method-keys}{\code{CascadingTheme$keys()}}
\item \href{#method-inheritance}{\code{CascadingTheme$inheritance()}}
\item \href{#method-value_validators}{\code{CascadingTheme$value_validators()}}
\item \href{#method-node_validators}{\code{CascadingTheme$node_validators()}}
\item \href{#method-node}{\code{CascadingTheme$node()}}
\item \href{#method-value}{\code{CascadingTheme$value()}}
\item \href{#method-value_base}{\code{CascadingTheme$value_base()}}
\item \href{#method-set_node}{\code{CascadingTheme$set_node()}}
\item \href{#method-set_value}{\code{CascadingTheme$set_value()}}
\item \href{#method-set_nodes}{\code{CascadingTheme$set_nodes()}}
\item \href{#method-set_values}{\code{CascadingTheme$set_values()}}
\item \href{#method-set_single_value_validators}{\code{CascadingTheme$set_single_value_validators()}}
\item \href{#method-set_single_node_validators}{\code{CascadingTheme$set_single_node_validators()}}
\item \href{#method-set_tree}{\code{CascadingTheme$set_tree()}}
\item \href{#method-set_data}{\code{CascadingTheme$set_data()}}
\item \href{#method-set_node_validators}{\code{CascadingTheme$set_node_validators()}}
\item \href{#method-set_value_validators}{\code{CascadingTheme$set_value_validators()}}
\item \href{#method-clone}{\code{CascadingTheme$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$new(data = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$compile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keys"></a>}}
\subsection{Method \code{keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inheritance"></a>}}
\subsection{Method \code{inheritance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$inheritance(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value_validators"></a>}}
\subsection{Method \code{value_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$value_validators(
  key,
  default = abort(glue::glue("No validators for value: '{key}'"))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-node_validators"></a>}}
\subsection{Method \code{node_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$node_validators(
  key,
  default = abort(glue::glue("No validators for node: '{key}'"))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-node"></a>}}
\subsection{Method \code{node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$node(key, default = abort(glue::glue("No such node: '{key}'")))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\subsection{Method \code{value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$value(
  key,
  default = abort(glue::glue("No such value: '{key}'"))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value_base"></a>}}
\subsection{Method \code{value_base()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$value_base(
  key,
  default = abort(glue::glue("No such node: '{key}'"))
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_node"></a>}}
\subsection{Method \code{set_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_node(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_value"></a>}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_value(key, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_nodes"></a>}}
\subsection{Method \code{set_nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_nodes(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_values"></a>}}
\subsection{Method \code{set_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_values(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_single_value_validators"></a>}}
\subsection{Method \code{set_single_value_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_single_value_validators(key, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_single_node_validators"></a>}}
\subsection{Method \code{set_single_node_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_single_node_validators(key, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_tree"></a>}}
\subsection{Method \code{set_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_tree(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_data(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_node_validators"></a>}}
\subsection{Method \code{set_node_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_node_validators(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_value_validators"></a>}}
\subsection{Method \code{set_value_validators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$set_value_validators(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CascadingTheme$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
