% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-mapping.R
\name{ColumnMapping}
\alias{ColumnMapping}
\title{Create a tibble from user data}
\description{
Create a tibble from user data

Create a tibble from user data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-aesthetics}{\code{ColumnMapping$aesthetics()}}
\item \href{#method-map}{\code{ColumnMapping$map()}}
\item \href{#method-map_new}{\code{ColumnMapping$map_new()}}
\item \href{#method-map_update}{\code{ColumnMapping$map_update()}}
\item \href{#method-clone}{\code{ColumnMapping$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aesthetics"></a>}}
\subsection{Method \code{aesthetics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMapping$aesthetics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMapping$map(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_new"></a>}}
\subsection{Method \code{map_new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMapping$map_new(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_update"></a>}}
\subsection{Method \code{map_update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMapping$map_update(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColumnMapping$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
