% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\docType{data}
\name{Coord}
\alias{Coord}
\alias{CoordIdentity}
\title{Coordinate systems}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Coord

CoordIdentity
}
\description{
Coordinate systems pick which \code{aesthetics} are position \code{aesthetics},
finalize the \link{Panel} at the end of the build process, and transform
coordinates. Regardless of their \code{aesthetics}, they should always
return from their \code{$transform()} method a data frame with \code{x} and \code{y}
columns. The default \code{CoordIdentity} does not change the \link{Panel} and
performs no changes in its \code{$transform()} method.
}
\section{Coord R6 Methods}{


\code{Coord$new(aesthetics = c("x", "y"))}
\itemize{
\item Create a coordinate system representing one or more position \code{aesthetics}.
}

\code{coord$transform(data_mapped, scales)}
\itemize{
\item Transform a data frame containing columns representing one or more position \code{$aesthetics}. This method should always return a data frame with \code{x} and \code{y} columns and  leave non-position columns unchanged.
}

\code{coord$finalize_panel(panel)}
\itemize{
\item Finalize the \link{Panel} object at the end of the build process. This is where scale expansion (if any) can occur.
}
}

\section{CoordIdentity inheritance}{
 CoordIdentity inherits from \code{\link{Coord}}
}

\keyword{datasets}
