% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\docType{data}
\name{Facet}
\alias{Facet}
\alias{FacetNull}
\title{Split data in to one or more panels}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Facet

FacetNull
}
\description{
Facets as implemented in ggr6 split layer data into
one or more pieces, and generate the list of \link{Panel}
objects that contain the list of \link{Scale}s, the list of
\link{Guide}s, and a reference to the \link{Coord}inate system.
Eventually they will have to also draw the panels (currently
this is the job of the \link{Renderer}). The default \code{FacetNull}
draws one panel with all of each \link{Layer}'s data.
}
\section{FacetNull inheritance}{
 FacetNull inherits from \code{\link{Facet}}
}

\keyword{datasets}
