% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{Geom}
\alias{Geom}
\alias{GeomBlank}
\title{Render a graphical object}
\description{
Render a graphical object

Render a graphical object
}
\section{Super class}{
\code{\link[ggr6:StatGeomBase]{ggr6::StatGeomBase}} -> \code{Geom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render_panel}{\code{Geom$render_panel()}}
\item \href{#method-render_group}{\code{Geom$render_group()}}
\item \href{#method-clone}{\code{Geom$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="aesthetics">}\href{../../ggr6/html/StatGeomBase.html#method-aesthetics}{\code{ggr6::StatGeomBase$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_constant_aesthetic_values}{\code{ggr6::StatGeomBase$apply_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_default_aesthetic_values}{\code{ggr6::StatGeomBase$apply_default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-default_aesthetic_values}{\code{ggr6::StatGeomBase$default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="initialize">}\href{../../ggr6/html/StatGeomBase.html#method-initialize}{\code{ggr6::StatGeomBase$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="set_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-set_constant_aesthetic_values}{\code{ggr6::StatGeomBase$set_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="setup_data">}\href{../../ggr6/html/StatGeomBase.html#method-setup_data}{\code{ggr6::StatGeomBase$setup_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panel"></a>}}
\subsection{Method \code{render_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Geom$render_panel(data, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_group"></a>}}
\subsection{Method \code{render_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Geom$render_group(data, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Geom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[ggr6:StatGeomBase]{ggr6::StatGeomBase}} -> \code{\link[ggr6:Geom]{ggr6::Geom}} -> \code{GeomBlank}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render_panel}{\code{GeomBlank$render_panel()}}
\item \href{#method-clone}{\code{GeomBlank$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="aesthetics">}\href{../../ggr6/html/StatGeomBase.html#method-aesthetics}{\code{ggr6::StatGeomBase$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_constant_aesthetic_values}{\code{ggr6::StatGeomBase$apply_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_default_aesthetic_values}{\code{ggr6::StatGeomBase$apply_default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-default_aesthetic_values}{\code{ggr6::StatGeomBase$default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="initialize">}\href{../../ggr6/html/StatGeomBase.html#method-initialize}{\code{ggr6::StatGeomBase$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="set_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-set_constant_aesthetic_values}{\code{ggr6::StatGeomBase$set_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="setup_data">}\href{../../ggr6/html/StatGeomBase.html#method-setup_data}{\code{ggr6::StatGeomBase$setup_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Geom" data-id="render_group">}\href{../../ggr6/html/Geom.html#method-render_group}{\code{ggr6::Geom$render_group()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panel"></a>}}
\subsection{Method \code{render_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBlank$render_panel(data, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeomBlank$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
