% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggr6-package.R
\docType{package}
\name{ggr6-package}
\alias{ggr6}
\alias{ggr6-package}
\title{ggr6: An Implementaation of the Grammar of Graphics in R6}
\description{
Implements the Grammar of Graphics using R6 reference objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paleolimbot.github.io/ggr6}
  \item \url{https://github.com/paleolimbot/ggr6}
  \item Report bugs at \url{https://github.com/paleolimbot/ggr6/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

}
\keyword{internal}
