% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic.R
\name{Graphic}
\alias{Graphic}
\title{Specify a Graphic}
\description{
Specify a Graphic

Specify a Graphic
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Graphic$new()}}
\item \href{#method-add_layer}{\code{Graphic$add_layer()}}
\item \href{#method-add_scale}{\code{Graphic$add_scale()}}
\item \href{#method-replace_scale}{\code{Graphic$replace_scale()}}
\item \href{#method-set_coord}{\code{Graphic$set_coord()}}
\item \href{#method-set_facet}{\code{Graphic$set_facet()}}
\item \href{#method-theme_set}{\code{Graphic$theme_set()}}
\item \href{#method-set_theme}{\code{Graphic$set_theme()}}
\item \href{#method-clone}{\code{Graphic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_layer"></a>}}
\subsection{Method \code{add_layer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$add_layer(layer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_scale"></a>}}
\subsection{Method \code{add_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$add_scale(scale)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_scale"></a>}}
\subsection{Method \code{replace_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$replace_scale(scale)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_coord"></a>}}
\subsection{Method \code{set_coord()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$set_coord(coord)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_facet"></a>}}
\subsection{Method \code{set_facet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$set_facet(facet)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-theme_set"></a>}}
\subsection{Method \code{theme_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$theme_set(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_theme"></a>}}
\subsection{Method \code{set_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$set_theme(theme)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graphic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
