% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics-renderer.R
\name{GraphicsRenderer}
\alias{GraphicsRenderer}
\title{Graphics Renderer}
\description{
A \link{Renderer} based on the \link[graphics:graphics-package]{graphics} package. This
is mostly used for testing at the moment.
}
\section{Super class}{
\code{\link[ggr6:Renderer]{ggr6::Renderer}} -> \code{GraphicsRenderer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render_points}{\code{GraphicsRenderer$render_points()}}
\item \href{#method-render_path}{\code{GraphicsRenderer$render_path()}}
\item \href{#method-render_polygon}{\code{GraphicsRenderer$render_polygon()}}
\item \href{#method-render_text}{\code{GraphicsRenderer$render_text()}}
\item \href{#method-render_null}{\code{GraphicsRenderer$render_null()}}
\item \href{#method-render_stack}{\code{GraphicsRenderer$render_stack()}}
\item \href{#method-render_panel}{\code{GraphicsRenderer$render_panel()}}
\item \href{#method-render_panels}{\code{GraphicsRenderer$render_panels()}}
\item \href{#method-default_scale}{\code{GraphicsRenderer$default_scale()}}
\item \href{#method-clone}{\code{GraphicsRenderer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Renderer" data-id="aesthetics">}\href{../../ggr6/html/Renderer.html#method-aesthetics}{\code{ggr6::Renderer$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Renderer" data-id="default_aesthetics">}\href{../../ggr6/html/Renderer.html#method-default_aesthetics}{\code{ggr6::Renderer$default_aesthetics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_points"></a>}}
\subsection{Method \code{render_points()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_points(
  x,
  y,
  feature = NULL,
  pch = theme("points.pch", 16),
  cex = theme("points.cex", 1),
  col = theme("points.col", "black"),
  lwd = theme("points.lwd", 1),
  bg = theme("points.bg", "black"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_path"></a>}}
\subsection{Method \code{render_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_path(
  x,
  y,
  feature = 1L,
  part = 1L,
  lty = theme("path.lty", 1),
  lwd = theme("path.lwd", 1),
  col = theme("path.col", "black"),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_polygon"></a>}}
\subsection{Method \code{render_polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_polygon(
  x,
  y,
  feature = 1L,
  piece = 1L,
  part = 1L,
  col = theme("polygon.col", "black"),
  fill = theme("polygon.fill", "grey50"),
  lty = theme("polygon.lty", 1),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_text"></a>}}
\subsection{Method \code{render_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_text(
  x,
  y,
  label,
  feature = 1L,
  cex = theme("text.cex", 1),
  col = theme("text.col", "black"),
  font = theme("text.font", 1),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_null"></a>}}
\subsection{Method \code{render_null()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_null()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_stack"></a>}}
\subsection{Method \code{render_stack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_stack(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panel"></a>}}
\subsection{Method \code{render_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_panel(panel, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panels"></a>}}
\subsection{Method \code{render_panels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$render_panels(graphic, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_scale"></a>}}
\subsection{Method \code{default_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$default_scale(x, aesthetic)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GraphicsRenderer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
