% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\docType{data}
\name{GuideList}
\alias{GuideList}
\title{List of Guides}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GuideList
}
\description{
List of Guides
}
\section{GuideList R6 Methods}{


\code{guide_list$guide(aesthetic, default = NULL)} (inherited from \code{\link{GuideList}})
\itemize{
\item Extract a guide by \code{aesthetic} (returning \code{default} if there is none.)
}

\code{guide_list$train_layers(layers, renderer)}
\itemize{
\item Calls the \code{$train_layers()} method of each \link{Guide}.
}

\code{guide_list$merge_all()}
\itemize{
\item Merges the \link{Guide} objects by calling their \code{$merge()} methods iteratively. Merging is done in-place.
}
}

\keyword{datasets}
