% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascading-theme.R
\name{is_cascading_element}
\alias{is_cascading_element}
\alias{new_cascading_element}
\alias{validate_cascading_element}
\title{S3 details for cascading elements}
\usage{
is_cascading_element(x)

new_cascading_element(x, subclass = NULL)

validate_cascading_element(x)
}
\arguments{
\item{x}{(possibly) a \code{\link[=cascading_element]{cascading_element()}}}

\item{subclass}{The class to use for this cascading element}
}
\description{
S3 details for cascading elements
}
\examples{
is_cascading_element(cascading_element())
new_cascading_element(list())
validate_cascading_element(cascading_element())

}
