% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{Layer}
\alias{Layer}
\title{A Graphics Layer}
\description{
Specifies a Layer within a \link{Graphic}. Each \link{Layer} in the
Grammar of Graphics framework contains a \link{Stat}, a \link{Geom}, and
a \link{Position}. The \code{data_src} and \link{ColumnMapping} are used to create
the data for each layer; the \code{stat_mapping} is a \link{ColumnMapping} that
is applied after the \link{Stat} is computed.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Layer$new()}}
\item \href{#method-data}{\code{Layer$data()}}
\item \href{#method-data_stat}{\code{Layer$data_stat()}}
\item \href{#method-clone}{\code{Layer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$new(
  data_src = NULL,
  mapping = ColumnMappingIdentity$new(data_src),
  stat_mapping = ColumnMappingIdentity$new(),
  geom = GeomBlank$new(),
  stat = StatIdentity$new(),
  position = PositionIdentity$new()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$data(data_src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_stat"></a>}}
\subsection{Method \code{data_stat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$data_stat(data_stat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
