% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\docType{data}
\name{LayerList}
\alias{LayerList}
\title{A list of Layers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LayerList
}
\description{
A mutable list of \link{Layer} objects.
}
\keyword{datasets}
