% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascading-theme.R
\name{missing_value}
\alias{missing_value}
\alias{is_missing_value}
\title{Sentinel for missing values}
\usage{
missing_value()

is_missing_value(x)
}
\arguments{
\item{x}{A (possibly) missing value}
}
\description{
Sentinel for missing values
}
