% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\docType{data}
\name{Panel}
\alias{Panel}
\title{Panels}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Panel
}
\description{
Panel objects are the \link{Geom}-, \link{Stat}-, and \link{Position}-facing
representation of the \link{Coord}, \link{Scale}, and \link{Guide} objects.
Panels are created by the \link{Facet} and finalized by the \link{Coord} at the
end of the build process. Currently they also train the \link{Guide} objects,
but this is against the philosphy of the Panel (a vessel in which information
relevant to the Panel is contained) and should be moved to the \link{Builder}.
}
\section{Panel R6 Methods}{


\code{Panel$new(coord = CoordIdentity$new(), scales = ScaleList$new())}
\itemize{
\item Create a Panel with a \link{Coord} and \link{ScaleList}.
}

\code{panel$train_guides(layers, renderer)}
\itemize{
\item This method should be moved to the \link{Builder}.
}

\code{panel$position_scales()}
\itemize{
\item Returns a \link{ScaleList} for each position aesthetic defined by the \link{Coord}.
}

\code{panel$non_position_scales()}
\itemize{
\item Returns a \link{ScaleList} of scales representing non-position aesthetics.
}

\code{panel$transform(data)}
\itemize{
\item Calls the \link{Coord}'s \code{$transform()} method.
}
}

\keyword{datasets}
