% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\docType{data}
\name{Position}
\alias{Position}
\alias{PositionIdentity}
\title{Position adjustment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Position

PositionIdentity
}
\description{
Position adjustment
}
\keyword{datasets}
