% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\docType{data}
\name{Renderer}
\alias{Renderer}
\alias{RendererIdentity}
\title{Render and assemble a Graphic}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Renderer

RendererIdentity
}
\description{
The role of the renderer isn't solidified in ggr6. Currently it's an
abstraction of various graphics backend that allow ggr6 to focus on the
grammar (rather than on rendering graphical objects). Currently it has
methods for rendering graphical primitives, methods for finding
the aesthetics that are supported by each primitive, methods for
constructing default scales for each aesthetic, and methods for
assembling panels.
}
\details{
Realistically, each renderer needs its own collection
of \link{Geom}, \link{Facet}, \link{Coord}, and \link{Scale} subclasses.
The role of ggr6 is to provide robust superclasses for these such
that the subclasses can focus on rendering. This package provides
\code{RendererIdentity} for testing and \link{RendererGraphics}, which uses the
base \link[graphics:graphics-package]{graphics} package to render graphics.
}
\section{RendererIdentity inheritance}{
 RendererIdentity inherits from \code{\link{Renderer}}
}

\keyword{datasets}
