% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer-graphics.R
\docType{data}
\name{RendererGraphics}
\alias{RendererGraphics}
\title{Graphics Renderer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RendererGraphics
}
\description{
A \link{Renderer} based on the \link[graphics:graphics-package]{graphics} package. This
is mostly used for testing at the moment.
}
\section{RendererGraphics inheritance}{
 RendererGraphics inherits from \code{\link{Renderer}}
}

\keyword{datasets}
