% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale_none}
\alias{rescale_none}
\title{Don't rescale}
\usage{
rescale_none(x, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{...}{Used to accept any other parameters.}
}
\value{
\code{x} unchanged
}
\description{
This rescaler just returns its input, as opposed to
\code{\link[scales:rescale]{scales::rescale()}}, \code{\link[scales:rescale_max]{scales::rescale_max()}}, and \code{\link[scales:rescale_mid]{scales::rescale_mid()}},
all of which return values between 0 and 1.
}
\examples{
scales::rescale(1:5)
rescale_none(1:5)

}
