% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\docType{data}
\name{Scale}
\alias{Scale}
\title{Convert between data spaces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Scale
}
\arguments{
\item{aesthetics}{A character vector of aesthetics that this
scale will represent.}
}
\description{
Scales are...
}
\keyword{datasets}
