% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-continuous.R
\name{ScaleContinuous}
\alias{ScaleContinuous}
\title{Continuous Scale}
\description{
Continuous Scale

Continuous Scale
}
\section{Super classes}{
\code{\link[ggr6:Scale]{ggr6::Scale}} -> \code{\link[ggr6:ScaleSimple]{ggr6::ScaleSimple}} -> \code{ScaleContinuous}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScaleContinuous$new()}}
\item \href{#method-map}{\code{ScaleContinuous$map()}}
\item \href{#method-within_limits}{\code{ScaleContinuous$within_limits()}}
\item \href{#method-set_trans}{\code{ScaleContinuous$set_trans()}}
\item \href{#method-set_palette}{\code{ScaleContinuous$set_palette()}}
\item \href{#method-clone}{\code{ScaleContinuous$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="name">}\href{../../ggr6/html/Scale.html#method-name}{\code{ggr6::Scale$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="set_guide">}\href{../../ggr6/html/Scale.html#method-set_guide}{\code{ggr6::Scale$set_guide()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="breaks">}\href{../../ggr6/html/ScaleSimple.html#method-breaks}{\code{ggr6::ScaleSimple$breaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="breaks_minor">}\href{../../ggr6/html/ScaleSimple.html#method-breaks_minor}{\code{ggr6::ScaleSimple$breaks_minor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="is_empty">}\href{../../ggr6/html/ScaleSimple.html#method-is_empty}{\code{ggr6::ScaleSimple$is_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="labels">}\href{../../ggr6/html/ScaleSimple.html#method-labels}{\code{ggr6::ScaleSimple$labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="limits">}\href{../../ggr6/html/ScaleSimple.html#method-limits}{\code{ggr6::ScaleSimple$limits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="map_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-map_tbl}{\code{ggr6::ScaleSimple$map_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="reset">}\href{../../ggr6/html/ScaleSimple.html#method-reset}{\code{ggr6::ScaleSimple$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_breaks">}\href{../../ggr6/html/ScaleSimple.html#method-set_breaks}{\code{ggr6::ScaleSimple$set_breaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_breaks_minor">}\href{../../ggr6/html/ScaleSimple.html#method-set_breaks_minor}{\code{ggr6::ScaleSimple$set_breaks_minor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_labels">}\href{../../ggr6/html/ScaleSimple.html#method-set_labels}{\code{ggr6::ScaleSimple$set_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_limits">}\href{../../ggr6/html/ScaleSimple.html#method-set_limits}{\code{ggr6::ScaleSimple$set_limits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_limits_empty">}\href{../../ggr6/html/ScaleSimple.html#method-set_limits_empty}{\code{ggr6::ScaleSimple$set_limits_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_na_value">}\href{../../ggr6/html/ScaleSimple.html#method-set_na_value}{\code{ggr6::ScaleSimple$set_na_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_oob">}\href{../../ggr6/html/ScaleSimple.html#method-set_oob}{\code{ggr6::ScaleSimple$set_oob()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_range">}\href{../../ggr6/html/ScaleSimple.html#method-set_range}{\code{ggr6::ScaleSimple$set_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_rescaler">}\href{../../ggr6/html/ScaleSimple.html#method-set_rescaler}{\code{ggr6::ScaleSimple$set_rescaler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="train">}\href{../../ggr6/html/ScaleSimple.html#method-train}{\code{ggr6::ScaleSimple$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="train_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-train_tbl}{\code{ggr6::ScaleSimple$train_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="trained_range">}\href{../../ggr6/html/ScaleSimple.html#method-trained_range}{\code{ggr6::ScaleSimple$trained_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="transform">}\href{../../ggr6/html/ScaleSimple.html#method-transform}{\code{ggr6::ScaleSimple$transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="transform_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-transform_tbl}{\code{ggr6::ScaleSimple$transform_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="untransform">}\href{../../ggr6/html/ScaleSimple.html#method-untransform}{\code{ggr6::ScaleSimple$untransform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="untransform_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-untransform_tbl}{\code{ggr6::ScaleSimple$untransform_tbl()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$new(aesthetics = character(0))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$map(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-within_limits"></a>}}
\subsection{Method \code{within_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$within_limits(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_trans"></a>}}
\subsection{Method \code{set_trans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$set_trans(trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_palette"></a>}}
\subsection{Method \code{set_palette()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$set_palette(palette)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleContinuous$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
