% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-discrete.R
\name{ScaleDiscrete}
\alias{ScaleDiscrete}
\title{Discrete Scale}
\description{
Discrete Scale

Discrete Scale
}
\section{Super classes}{
\code{\link[ggr6:Scale]{ggr6::Scale}} -> \code{\link[ggr6:ScaleSimple]{ggr6::ScaleSimple}} -> \code{ScaleDiscrete}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScaleDiscrete$new()}}
\item \href{#method-train}{\code{ScaleDiscrete$train()}}
\item \href{#method-palette}{\code{ScaleDiscrete$palette()}}
\item \href{#method-map}{\code{ScaleDiscrete$map()}}
\item \href{#method-within_limits}{\code{ScaleDiscrete$within_limits()}}
\item \href{#method-set_palette_factory}{\code{ScaleDiscrete$set_palette_factory()}}
\item \href{#method-set_drop}{\code{ScaleDiscrete$set_drop()}}
\item \href{#method-clone}{\code{ScaleDiscrete$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="name">}\href{../../ggr6/html/Scale.html#method-name}{\code{ggr6::Scale$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="set_guide">}\href{../../ggr6/html/Scale.html#method-set_guide}{\code{ggr6::Scale$set_guide()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="breaks">}\href{../../ggr6/html/ScaleSimple.html#method-breaks}{\code{ggr6::ScaleSimple$breaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="breaks_minor">}\href{../../ggr6/html/ScaleSimple.html#method-breaks_minor}{\code{ggr6::ScaleSimple$breaks_minor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="is_empty">}\href{../../ggr6/html/ScaleSimple.html#method-is_empty}{\code{ggr6::ScaleSimple$is_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="labels">}\href{../../ggr6/html/ScaleSimple.html#method-labels}{\code{ggr6::ScaleSimple$labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="limits">}\href{../../ggr6/html/ScaleSimple.html#method-limits}{\code{ggr6::ScaleSimple$limits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="map_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-map_tbl}{\code{ggr6::ScaleSimple$map_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="reset">}\href{../../ggr6/html/ScaleSimple.html#method-reset}{\code{ggr6::ScaleSimple$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_breaks">}\href{../../ggr6/html/ScaleSimple.html#method-set_breaks}{\code{ggr6::ScaleSimple$set_breaks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_breaks_minor">}\href{../../ggr6/html/ScaleSimple.html#method-set_breaks_minor}{\code{ggr6::ScaleSimple$set_breaks_minor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_labels">}\href{../../ggr6/html/ScaleSimple.html#method-set_labels}{\code{ggr6::ScaleSimple$set_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_limits">}\href{../../ggr6/html/ScaleSimple.html#method-set_limits}{\code{ggr6::ScaleSimple$set_limits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_limits_empty">}\href{../../ggr6/html/ScaleSimple.html#method-set_limits_empty}{\code{ggr6::ScaleSimple$set_limits_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_na_value">}\href{../../ggr6/html/ScaleSimple.html#method-set_na_value}{\code{ggr6::ScaleSimple$set_na_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_oob">}\href{../../ggr6/html/ScaleSimple.html#method-set_oob}{\code{ggr6::ScaleSimple$set_oob()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_range">}\href{../../ggr6/html/ScaleSimple.html#method-set_range}{\code{ggr6::ScaleSimple$set_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_rescaler">}\href{../../ggr6/html/ScaleSimple.html#method-set_rescaler}{\code{ggr6::ScaleSimple$set_rescaler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="set_trans">}\href{../../ggr6/html/ScaleSimple.html#method-set_trans}{\code{ggr6::ScaleSimple$set_trans()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="train_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-train_tbl}{\code{ggr6::ScaleSimple$train_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="trained_range">}\href{../../ggr6/html/ScaleSimple.html#method-trained_range}{\code{ggr6::ScaleSimple$trained_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="transform">}\href{../../ggr6/html/ScaleSimple.html#method-transform}{\code{ggr6::ScaleSimple$transform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="transform_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-transform_tbl}{\code{ggr6::ScaleSimple$transform_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="untransform">}\href{../../ggr6/html/ScaleSimple.html#method-untransform}{\code{ggr6::ScaleSimple$untransform()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="ScaleSimple" data-id="untransform_tbl">}\href{../../ggr6/html/ScaleSimple.html#method-untransform_tbl}{\code{ggr6::ScaleSimple$untransform_tbl()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$new(aesthetics = character(0))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$train(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-palette"></a>}}
\subsection{Method \code{palette()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$palette(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$map(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-within_limits"></a>}}
\subsection{Method \code{within_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$within_limits(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_palette_factory"></a>}}
\subsection{Method \code{set_palette_factory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$set_palette_factory(palette_factory)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_drop"></a>}}
\subsection{Method \code{set_drop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$set_drop(drop)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleDiscrete$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
