% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\docType{data}
\name{ScaleList}
\alias{ScaleList}
\title{Mutable List of Scale objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScaleList
}
\description{
ScaleList objects contain a list of \link{Scale}s, and have
methods that can perform the same operation on all the
scales they contain. They currently also generate the
\link{GuideList} of trained \link{Guide}s based on the \link{Scale}s
they contain.
}
\section{ScaleList R6 Methods}{


\code{scale_list$aesthetics()} (inherited from \code{\link{ScaleList}})
\itemize{
\item A character vector of aesthetics represented by the \link{Scale}s within the list.
}

\code{scale_list$scale(aesthetic, default = NULL)}
\itemize{
\item Extract a scale by \code{aesthetic} (returning \code{default} if there is none.)
}

\code{scale_list$guides()}
\itemize{
\item Returns a \link{GuideList} of \link{Guide}s extracted and trained from the \link{Scale}s within the object.
}

\code{scale_list$filter_by_aesthetics(aesthetics)}
\itemize{
\item Return a new \code{ScaleList} containing any scale representing one or more \code{aesthetics} specified.
}

\code{scale_list$discard_by_aesthetics(aesthetics)}
\itemize{
\item Return a new \code{ScaleList} containing any scale that does not represent the \code{aesthetics} specified.
}

\code{scale_list$transform_tbl(data)}
\itemize{
\item Calls the \code{$transform_tbl()} method for each \link{Scale} iteratively, returning the result.
}

\code{scale_list$untransform_tbl(data)}
\itemize{
\item Calls the \code{$untransform_tbl()} method for each \link{Scale} iteratively, returning the result.
}

\code{scale_list$train_tbl(data_trans)}
\itemize{
\item Calls the \code{$transform_tbl()} method for each \link{Scale}.
}

\code{scale_list$reset()}
\itemize{
\item Calls the \code{$reset()} method for each \link{Scale} iteratively, returning the result.
}

\code{scale_list$map_tbl(data_trans)}
\itemize{
\item Calls the \code{$map_tbl()} method for each \link{Scale} iteratively, returning the result.
}

\code{scale_list$add_missing(data, renderer)}
\itemize{
\item Add missing scales to this object based on data (whose names are aesthetics) and the \link{Renderer}.
}
}

\keyword{datasets}
