% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{Stat}
\alias{Stat}
\alias{StatIdentity}
\title{Transform data}
\description{
Transform data

Transform data
}
\section{Super class}{
\code{\link[ggr6:StatGeomBase]{ggr6::StatGeomBase}} -> \code{Stat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Stat$new()}}
\item \href{#method-set_default_mapping}{\code{Stat$set_default_mapping()}}
\item \href{#method-compute_panel}{\code{Stat$compute_panel()}}
\item \href{#method-compute_group}{\code{Stat$compute_group()}}
\item \href{#method-clone}{\code{Stat$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="aesthetics">}\href{../../ggr6/html/StatGeomBase.html#method-aesthetics}{\code{ggr6::StatGeomBase$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_constant_aesthetic_values}{\code{ggr6::StatGeomBase$apply_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_default_aesthetic_values}{\code{ggr6::StatGeomBase$apply_default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-default_aesthetic_values}{\code{ggr6::StatGeomBase$default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="set_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-set_constant_aesthetic_values}{\code{ggr6::StatGeomBase$set_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="setup_data">}\href{../../ggr6/html/StatGeomBase.html#method-setup_data}{\code{ggr6::StatGeomBase$setup_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$new(constant_aesthetic_values = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_default_mapping"></a>}}
\subsection{Method \code{set_default_mapping()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$set_default_mapping(default_mapping)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_panel"></a>}}
\subsection{Method \code{compute_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$compute_panel(data_trans, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_group"></a>}}
\subsection{Method \code{compute_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$compute_group(data_trans, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[ggr6:StatGeomBase]{ggr6::StatGeomBase}} -> \code{\link[ggr6:Stat]{ggr6::Stat}} -> \code{StatIdentity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compute_panel}{\code{StatIdentity$compute_panel()}}
\item \href{#method-clone}{\code{StatIdentity$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="aesthetics">}\href{../../ggr6/html/StatGeomBase.html#method-aesthetics}{\code{ggr6::StatGeomBase$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_constant_aesthetic_values}{\code{ggr6::StatGeomBase$apply_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="apply_default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-apply_default_aesthetic_values}{\code{ggr6::StatGeomBase$apply_default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="default_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-default_aesthetic_values}{\code{ggr6::StatGeomBase$default_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="set_constant_aesthetic_values">}\href{../../ggr6/html/StatGeomBase.html#method-set_constant_aesthetic_values}{\code{ggr6::StatGeomBase$set_constant_aesthetic_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="StatGeomBase" data-id="setup_data">}\href{../../ggr6/html/StatGeomBase.html#method-setup_data}{\code{ggr6::StatGeomBase$setup_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Stat" data-id="compute_group">}\href{../../ggr6/html/Stat.html#method-compute_group}{\code{ggr6::Stat$compute_group()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Stat" data-id="initialize">}\href{../../ggr6/html/Stat.html#method-initialize}{\code{ggr6::Stat$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Stat" data-id="set_default_mapping">}\href{../../ggr6/html/Stat.html#method-set_default_mapping}{\code{ggr6::Stat$set_default_mapping()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_panel"></a>}}
\subsection{Method \code{compute_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatIdentity$compute_panel(data_trans, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatIdentity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
