% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-geom-base.R
\name{StatGeomBase}
\alias{StatGeomBase}
\title{Base class for statistics and geometries}
\description{
Base class for statistics and geometries

Base class for statistics and geometries
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StatGeomBase$new()}}
\item \href{#method-set_constant_aesthetic_values}{\code{StatGeomBase$set_constant_aesthetic_values()}}
\item \href{#method-setup_data}{\code{StatGeomBase$setup_data()}}
\item \href{#method-aesthetics}{\code{StatGeomBase$aesthetics()}}
\item \href{#method-default_aesthetic_values}{\code{StatGeomBase$default_aesthetic_values()}}
\item \href{#method-apply_default_aesthetic_values}{\code{StatGeomBase$apply_default_aesthetic_values()}}
\item \href{#method-apply_constant_aesthetic_values}{\code{StatGeomBase$apply_constant_aesthetic_values()}}
\item \href{#method-clone}{\code{StatGeomBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$new(constant_aesthetic_values = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_constant_aesthetic_values"></a>}}
\subsection{Method \code{set_constant_aesthetic_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$set_constant_aesthetic_values(constant_aesthetic_values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup_data"></a>}}
\subsection{Method \code{setup_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$setup_data(data, panel, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aesthetics"></a>}}
\subsection{Method \code{aesthetics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$aesthetics(renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_aesthetic_values"></a>}}
\subsection{Method \code{default_aesthetic_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$default_aesthetic_values(renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_default_aesthetic_values"></a>}}
\subsection{Method \code{apply_default_aesthetic_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$apply_default_aesthetic_values(data, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_constant_aesthetic_values"></a>}}
\subsection{Method \code{apply_constant_aesthetic_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$apply_constant_aesthetic_values(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatGeomBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
