% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trans_discrete_new}
\alias{trans_discrete_new}
\alias{discrete_identity_trans}
\title{Discrete Transform}
\usage{
trans_discrete_new(
  ...,
  breaks = identity,
  minor_breaks = function(breaks, limits, n) NULL,
  format = identity,
  domain = c(-Inf, Inf)
)

discrete_identity_trans()
}
\arguments{
\item{..., breaks, minor_breaks, format, domain}{See \code{\link[scales:trans_new]{scales::trans_new()}}.}
}
\description{
Discrete transforms have different defaults than those used by
\code{\link[scales:trans_new]{scales::trans_new()}}. \code{discrete_identity_trans()} is probably
the only useful discrete transform.
}
