% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column-mapping.R
\docType{data}
\name{ColumnMapping}
\alias{ColumnMapping}
\title{Create a tibble from user data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ColumnMapping
}
\description{
Create a tibble from user data
}
\keyword{datasets}
