% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{Coord}
\alias{Coord}
\alias{CoordIdentity}
\title{Coordinate systems}
\description{
Coordinate systems pick which \code{aesthetics} are position \code{aesthetics},
finalize the \link{Panel} at the end of the build process, and transform
coordinates. Regardless of their \code{aesthetics}, they should always
return from their \verb{$transform()} method a data frame with \code{x} and \code{y}
columns. The default \code{CoordIdentity} does not change the \link{Panel} and
performs no changes in its \verb{$transform()} method.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Coord$new()}}
\item \href{#method-transform}{\code{Coord$transform()}}
\item \href{#method-finalize_panel}{\code{Coord$finalize_panel()}}
\item \href{#method-clone}{\code{Coord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$new(aesthetics = c("x", "y"))}\if{html}{\out{</div>}}
}

\subsection{Details}{
Create a coordinate system representing one or more position
\code{aesthetics}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$transform(data_mapped, scales)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Transform a data frame containing columns representing one or more
position \verb{$aesthetics}. This method should always return a data frame
with \code{x} and \code{y} columns and  leave non-position columns unchanged.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize_panel"></a>}}
\subsection{Method \code{finalize_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$finalize_panel(panel)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Finalize the \link{Panel} object at the end of the build process. This is
where scale expansion (if any) can occur.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ggr6:Coord]{ggr6::Coord}} -> \code{CoordIdentity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-transform}{\code{CoordIdentity$transform()}}
\item \href{#method-finalize_panel}{\code{CoordIdentity$finalize_panel()}}
\item \href{#method-clone}{\code{CoordIdentity$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Coord" data-id="initialize">}\href{../../ggr6/html/Coord.html#method-initialize}{\code{ggr6::Coord$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordIdentity$transform(data_mapped, scales)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize_panel"></a>}}
\subsection{Method \code{finalize_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordIdentity$finalize_panel(panel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoordIdentity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
