% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{Facet}
\alias{Facet}
\alias{FacetNull}
\title{Split data in to one or more panels}
\description{
Facets as implemented in ggr6 split layer data into
one or more pieces, and generate the list of \link{Panel}
objects that contain the list of \link{Scale}s, the list of
\link{Guide}s, and a reference to the \link{Coord}inate system.
Eventually they will have to also draw the panels (currently
this is the job of the \link{Renderer}). The default \code{FacetNull}
draws one panel with all of each \link{Layer}'s data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-train}{\code{Facet$train()}}
\item \href{#method-reset}{\code{Facet$reset()}}
\item \href{#method-panels}{\code{Facet$panels()}}
\item \href{#method-panel_data}{\code{Facet$panel_data()}}
\item \href{#method-clone}{\code{Facet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Facet$train(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Facet$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-panels"></a>}}
\subsection{Method \code{panels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Facet$panels(coord, scales)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-panel_data"></a>}}
\subsection{Method \code{panel_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Facet$panel_data(data, panel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Facet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ggr6:Facet]{ggr6::Facet}} -> \code{FacetNull}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-train}{\code{FacetNull$train()}}
\item \href{#method-reset}{\code{FacetNull$reset()}}
\item \href{#method-panels}{\code{FacetNull$panels()}}
\item \href{#method-panel_data}{\code{FacetNull$panel_data()}}
\item \href{#method-clone}{\code{FacetNull$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetNull$train(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetNull$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-panels"></a>}}
\subsection{Method \code{panels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetNull$panels(coord, scales)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-panel_data"></a>}}
\subsection{Method \code{panel_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetNull$panel_data(data, panel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FacetNull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
