% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{Geom}
\alias{Geom}
\alias{GeomBlank}
\title{Render a graphical object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Geom

GeomBlank
}
\description{
Render a graphical object
}
\keyword{datasets}
