% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\docType{data}
\name{Guide}
\alias{Guide}
\alias{GuideNull}
\title{Graphical represntation of Scales}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Guide

GuideNull
}
\description{
Guide objects are a graphical representation of \code{\link{Scale}}s, in that
they communicate the relationship between user data values and
mapped data values. Each \code{\link{Scale}} must have a Guide, although this
can be (and often is) \code{GuideNull}, which draws nothing). Guides which
have the same name, class, breaks, and labels are merged together,
such that in the final plot there may be more than one scale represented
by one guide. All guides except \code{GuideNull} will probably have to be
subclassed for each \code{\link{Renderer}}.
}
\keyword{datasets}
