% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\docType{data}
\name{Layer}
\alias{Layer}
\title{A Graphics Layer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Layer
}
\description{
Specifies a Layer within a \link{Graphic}. Each \link{Layer} in the
Grammar of Graphics framework contains a \link{Stat}, a \link{Geom}, and
a \link{Position}. The \code{data_src} and \link{ColumnMapping} are used to create
the data for each layer; the \code{stat_mapping} is a \link{ColumnMapping} that
is applied after the \link{Stat} is computed.
}
\keyword{datasets}
