% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{LayerList}
\alias{LayerList}
\title{A list of Layers}
\description{
A mutable list of \link{Layer} objects.
}
\section{Super class}{
\code{\link[ggr6:List]{ggr6::List}} -> \code{LayerList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set}{\code{LayerList$set()}}
\item \href{#method-clone}{\code{LayerList$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="add">}\href{../../ggr6/html/List.html#method-add}{\code{ggr6::List$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="get">}\href{../../ggr6/html/List.html#method-get}{\code{ggr6::List$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="initialize">}\href{../../ggr6/html/List.html#method-initialize}{\code{ggr6::List$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="size">}\href{../../ggr6/html/List.html#method-size}{\code{ggr6::List$size()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerList$set(index, item)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
