% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{Panel}
\alias{Panel}
\title{Panels}
\description{
Panel objects are the \link{Geom}-, \link{Stat}-, and \link{Position}-facing
representation of the \link{Coord}, \link{Scale}, and \link{Guide} objects.
Panels are created by the \link{Facet} and finalized by the \link{Coord} at the
end of the build process. Currently they also train the \link{Guide} objects,
but this is against the philosphy of the Panel (a vessel in which information
relevant to the Panel is contained) and should be moved to the \link{Builder}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Panel$new()}}
\item \href{#method-train_guides}{\code{Panel$train_guides()}}
\item \href{#method-position_scales}{\code{Panel$position_scales()}}
\item \href{#method-non_position_scales}{\code{Panel$non_position_scales()}}
\item \href{#method-transform}{\code{Panel$transform()}}
\item \href{#method-clone}{\code{Panel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$new(coord = CoordIdentity$new(), scales = ScaleList$new())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_guides"></a>}}
\subsection{Method \code{train_guides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$train_guides(layers, renderer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-position_scales"></a>}}
\subsection{Method \code{position_scales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$position_scales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-non_position_scales"></a>}}
\subsection{Method \code{non_position_scales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$non_position_scales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$transform(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Panel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
