% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\docType{data}
\name{Position}
\alias{Position}
\alias{PositionIdentity}
\title{Adjust positions before rendering}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Position

PositionIdentity
}
\description{
Positions are called once per panel per layer between
the \link{Stat} computation and the \link{Geom} computation. The
default \code{PositionIdentity} does no position adjustment.
}
\section{Position R6 Methods}{


\code{position$compute_panel(transformed_data, panel)}
\itemize{
\item Perform position adjustment, returning a transformed copy of \code{transformed_data}.
}
}

\section{PositionIdentity inheritance}{
 PositionIdentity inherits from \code{\link{Position}}
}

\keyword{datasets}
