% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{Position}
\alias{Position}
\alias{PositionIdentity}
\title{Adjust positions before rendering}
\description{
Positions are called once per panel per layer between
the \link{Stat} computation and the \link{Geom} computation. The
default \code{PositionIdentity} does no position adjustment.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compute_panel}{\code{Position$compute_panel()}}
\item \href{#method-clone}{\code{Position$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_panel"></a>}}
\subsection{Method \code{compute_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$compute_panel(transformed_data, panel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Position$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ggr6:Position]{ggr6::Position}} -> \code{PositionIdentity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-compute_panel}{\code{PositionIdentity$compute_panel()}}
\item \href{#method-clone}{\code{PositionIdentity$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_panel"></a>}}
\subsection{Method \code{compute_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionIdentity$compute_panel(transformed_data, panel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PositionIdentity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
