% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\docType{data}
\name{Renderer}
\alias{Renderer}
\alias{RendererIdentity}
\title{Render and assemble a Graphic}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Renderer

RendererIdentity
}
\description{
Render and assemble a Graphic
}
\keyword{datasets}
