% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer-graphics.R
\name{RendererGraphics}
\alias{RendererGraphics}
\title{Graphics Renderer}
\description{
A \link{Renderer} based on the \link[graphics:graphics-package]{graphics} package. This
is mostly used for testing at the moment.
}
\section{Super class}{
\code{\link[ggr6:Renderer]{ggr6::Renderer}} -> \code{RendererGraphics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render_points}{\code{RendererGraphics$render_points()}}
\item \href{#method-render_path}{\code{RendererGraphics$render_path()}}
\item \href{#method-render_polygon}{\code{RendererGraphics$render_polygon()}}
\item \href{#method-render_text}{\code{RendererGraphics$render_text()}}
\item \href{#method-render_null}{\code{RendererGraphics$render_null()}}
\item \href{#method-render_stack}{\code{RendererGraphics$render_stack()}}
\item \href{#method-render_panel}{\code{RendererGraphics$render_panel()}}
\item \href{#method-render_panels}{\code{RendererGraphics$render_panels()}}
\item \href{#method-default_scale}{\code{RendererGraphics$default_scale()}}
\item \href{#method-clone}{\code{RendererGraphics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Renderer" data-id="aesthetics">}\href{../../ggr6/html/Renderer.html#method-aesthetics}{\code{ggr6::Renderer$aesthetics()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Renderer" data-id="default_aesthetics">}\href{../../ggr6/html/Renderer.html#method-default_aesthetics}{\code{ggr6::Renderer$default_aesthetics()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_points"></a>}}
\subsection{Method \code{render_points()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_points(
  x,
  y,
  pch = 16,
  cex = 1,
  col = "black",
  lwd = 1,
  bg = "black",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_path"></a>}}
\subsection{Method \code{render_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_path(
  x,
  y,
  group = 1,
  lty = 1,
  lwd = 1,
  col = "black",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_polygon"></a>}}
\subsection{Method \code{render_polygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_polygon(
  x,
  y,
  group = 1,
  subgroup = 1,
  col = "black",
  fill = "grey50",
  lty = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_text"></a>}}
\subsection{Method \code{render_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_text(
  x,
  y,
  label,
  cex = 1,
  col = "black",
  font = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_null"></a>}}
\subsection{Method \code{render_null()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_null()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_stack"></a>}}
\subsection{Method \code{render_stack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_stack(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panel"></a>}}
\subsection{Method \code{render_panel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_panel(panel, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_panels"></a>}}
\subsection{Method \code{render_panels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$render_panels(graphic, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-default_scale"></a>}}
\subsection{Method \code{default_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$default_scale(x, aesthetic)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RendererGraphics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
