% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{Scale}
\alias{Scale}
\alias{ScaleNull}
\title{Convert between data spaces}
\description{
Scales convert between the user data space, the data space used
to compute the \link{Stat}s (transformed data space), and the data space
used by the \link{Geom}s (mapped data space) to draw graphical objects.
Scales also compute breaks and labels used in the \link{Guide}s. Most useful
Scale classes inherit from \link{ScaleSimple}, which is designed to leverage
the \link[scales:scales]{scales} package. Each scale represents zero
or more \verb{$aesthetics} (character vector), and contains a \verb{$guide}
(\link{Guide} object) that will be used to communicate the relationship
between user data values and mapped data values. \code{ScaleNull} is a
scale that has infinite limits and never transforms or maps values,
which is useful for testing and for values that should be passed
to the \link{Stat} and \link{Geom} without being transformed or mapped.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Scale$new()}}
\item \href{#method-name}{\code{Scale$name()}}
\item \href{#method-transform}{\code{Scale$transform()}}
\item \href{#method-untransform}{\code{Scale$untransform()}}
\item \href{#method-train}{\code{Scale$train()}}
\item \href{#method-reset}{\code{Scale$reset()}}
\item \href{#method-map}{\code{Scale$map()}}
\item \href{#method-transform_tbl}{\code{Scale$transform_tbl()}}
\item \href{#method-untransform_tbl}{\code{Scale$untransform_tbl()}}
\item \href{#method-train_tbl}{\code{Scale$train_tbl()}}
\item \href{#method-map_tbl}{\code{Scale$map_tbl()}}
\item \href{#method-breaks}{\code{Scale$breaks()}}
\item \href{#method-breaks_minor}{\code{Scale$breaks_minor()}}
\item \href{#method-labels}{\code{Scale$labels()}}
\item \href{#method-limits}{\code{Scale$limits()}}
\item \href{#method-within_limits}{\code{Scale$within_limits()}}
\item \href{#method-set_guide}{\code{Scale$set_guide()}}
\item \href{#method-clone}{\code{Scale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$new(aesthetics = character(0))}\if{html}{\out{</div>}}
}

\subsection{Details}{
Create a Scale object representing zero or more \code{aesthetics}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$name(x)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Calculate the scale name, which should be a character vector of
length 1. May return \code{\link[=waiver]{waiver()}} for when
the scale name should be calculated from the \code{\link{ColumnMapping}}. This
method may be removed in favour of this value living in the
\link{Guide}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$transform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform"></a>}}
\subsection{Method \code{untransform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$untransform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$train(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$map(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_tbl"></a>}}
\subsection{Method \code{transform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$transform_tbl(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform_tbl"></a>}}
\subsection{Method \code{untransform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$untransform_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_tbl"></a>}}
\subsection{Method \code{train_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$train_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_tbl"></a>}}
\subsection{Method \code{map_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$map_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks"></a>}}
\subsection{Method \code{breaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$breaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks_minor"></a>}}
\subsection{Method \code{breaks_minor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$breaks_minor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-labels"></a>}}
\subsection{Method \code{labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-limits"></a>}}
\subsection{Method \code{limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$limits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-within_limits"></a>}}
\subsection{Method \code{within_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$within_limits(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_guide"></a>}}
\subsection{Method \code{set_guide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$set_guide(guide)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[ggr6:Scale]{ggr6::Scale}} -> \code{ScaleNull}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-transform}{\code{ScaleNull$transform()}}
\item \href{#method-untransform}{\code{ScaleNull$untransform()}}
\item \href{#method-train}{\code{ScaleNull$train()}}
\item \href{#method-reset}{\code{ScaleNull$reset()}}
\item \href{#method-map}{\code{ScaleNull$map()}}
\item \href{#method-transform_tbl}{\code{ScaleNull$transform_tbl()}}
\item \href{#method-untransform_tbl}{\code{ScaleNull$untransform_tbl()}}
\item \href{#method-train_tbl}{\code{ScaleNull$train_tbl()}}
\item \href{#method-map_tbl}{\code{ScaleNull$map_tbl()}}
\item \href{#method-breaks}{\code{ScaleNull$breaks()}}
\item \href{#method-breaks_minor}{\code{ScaleNull$breaks_minor()}}
\item \href{#method-labels}{\code{ScaleNull$labels()}}
\item \href{#method-limits}{\code{ScaleNull$limits()}}
\item \href{#method-within_limits}{\code{ScaleNull$within_limits()}}
\item \href{#method-clone}{\code{ScaleNull$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="initialize">}\href{../../ggr6/html/Scale.html#method-initialize}{\code{ggr6::Scale$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="name">}\href{../../ggr6/html/Scale.html#method-name}{\code{ggr6::Scale$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="set_guide">}\href{../../ggr6/html/Scale.html#method-set_guide}{\code{ggr6::Scale$set_guide()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$transform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform"></a>}}
\subsection{Method \code{untransform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$untransform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$train(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map"></a>}}
\subsection{Method \code{map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$map(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_tbl"></a>}}
\subsection{Method \code{transform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$transform_tbl(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform_tbl"></a>}}
\subsection{Method \code{untransform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$untransform_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_tbl"></a>}}
\subsection{Method \code{train_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$train_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_tbl"></a>}}
\subsection{Method \code{map_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$map_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks"></a>}}
\subsection{Method \code{breaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$breaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks_minor"></a>}}
\subsection{Method \code{breaks_minor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$breaks_minor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-labels"></a>}}
\subsection{Method \code{labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-limits"></a>}}
\subsection{Method \code{limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$limits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-within_limits"></a>}}
\subsection{Method \code{within_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$within_limits(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleNull$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
