% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{ScaleList}
\alias{ScaleList}
\title{Mutable List of Scale objects}
\description{
ScaleList objects contain a list of \link{Scale}s, and have
methods that can perform the same operation on all the
scales they contain. They currently also generate the
\link{GuideList} of trained \link{Guide}s based on the \link{Scale}s
they contain.
}
\section{Super class}{
\code{\link[ggr6:List]{ggr6::List}} -> \code{ScaleList}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set}{\code{ScaleList$set()}}
\item \href{#method-aesthetics}{\code{ScaleList$aesthetics()}}
\item \href{#method-scale}{\code{ScaleList$scale()}}
\item \href{#method-guides}{\code{ScaleList$guides()}}
\item \href{#method-filter_by_aesthetics}{\code{ScaleList$filter_by_aesthetics()}}
\item \href{#method-discard_by_aesthetics}{\code{ScaleList$discard_by_aesthetics()}}
\item \href{#method-transform_tbl}{\code{ScaleList$transform_tbl()}}
\item \href{#method-untransform_tbl}{\code{ScaleList$untransform_tbl()}}
\item \href{#method-train_tbl}{\code{ScaleList$train_tbl()}}
\item \href{#method-reset}{\code{ScaleList$reset()}}
\item \href{#method-map_tbl}{\code{ScaleList$map_tbl()}}
\item \href{#method-add_missing}{\code{ScaleList$add_missing()}}
\item \href{#method-clone}{\code{ScaleList$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="add">}\href{../../ggr6/html/List.html#method-add}{\code{ggr6::List$add()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="get">}\href{../../ggr6/html/List.html#method-get}{\code{ggr6::List$get()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="initialize">}\href{../../ggr6/html/List.html#method-initialize}{\code{ggr6::List$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="List" data-id="size">}\href{../../ggr6/html/List.html#method-size}{\code{ggr6::List$size()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$set(index, item)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aesthetics"></a>}}
\subsection{Method \code{aesthetics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$aesthetics()}\if{html}{\out{</div>}}
}

\subsection{Details}{
A character vector of aesthetics represented by the \link{Scale}s
within the list.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$scale(aesthetic, default = NULL)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Extract a scale by \code{aesthetic} (returning \code{default} if there is none.)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-guides"></a>}}
\subsection{Method \code{guides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$guides()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a \link{GuideList} of \link{Guide}s extracted and trained from the
\link{Scale}s within the object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_by_aesthetics"></a>}}
\subsection{Method \code{filter_by_aesthetics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$filter_by_aesthetics(aesthetics)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Return a new \code{ScaleList} containing any scale representing
one or more \code{aesthetics} specified.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discard_by_aesthetics"></a>}}
\subsection{Method \code{discard_by_aesthetics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$discard_by_aesthetics(aesthetics)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Return a new \code{ScaleList} containing any scale that does not
represent the \code{aesthetics} specified.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_tbl"></a>}}
\subsection{Method \code{transform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$transform_tbl(data)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Calls the \verb{$transform_tbl()} method for each \link{Scale} iteratively,
returning the result.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform_tbl"></a>}}
\subsection{Method \code{untransform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$untransform_tbl(data)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Calls the \verb{$untransform_tbl()} method for each \link{Scale} iteratively,
returning the result.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_tbl"></a>}}
\subsection{Method \code{train_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$train_tbl(data_trans)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Calls the \verb{$transform_tbl()} method for each \link{Scale}.
Calls the \verb{$reset()} method for each \link{Scale} iteratively,
returning the result.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_tbl"></a>}}
\subsection{Method \code{map_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$map_tbl(data_trans)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Calls the \verb{$map_tbl()} method for each \link{Scale} iteratively,
returning the result.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_missing"></a>}}
\subsection{Method \code{add_missing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$add_missing(data, renderer)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Add missing scales to this object based on data (whose names
are aesthetics) and the \link{Renderer}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
