% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-simple.R
\name{ScaleSimple}
\alias{ScaleSimple}
\title{Default implementation of Scale objects}
\description{
Most useful scales will inherit from \link{ScaleContinuous} or
\link{ScaleDiscrete}, both of which use the \link[scales:scales]{scales}
package to create meaningful mappings between user data and
aesthetics with useful breaks and labels.
}
\section{Super class}{
\code{\link[ggr6:Scale]{ggr6::Scale}} -> \code{ScaleSimple}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScaleSimple$new()}}
\item \href{#method-is_empty}{\code{ScaleSimple$is_empty()}}
\item \href{#method-transform}{\code{ScaleSimple$transform()}}
\item \href{#method-untransform}{\code{ScaleSimple$untransform()}}
\item \href{#method-transform_tbl}{\code{ScaleSimple$transform_tbl()}}
\item \href{#method-untransform_tbl}{\code{ScaleSimple$untransform_tbl()}}
\item \href{#method-train}{\code{ScaleSimple$train()}}
\item \href{#method-train_tbl}{\code{ScaleSimple$train_tbl()}}
\item \href{#method-reset}{\code{ScaleSimple$reset()}}
\item \href{#method-map_tbl}{\code{ScaleSimple$map_tbl()}}
\item \href{#method-breaks}{\code{ScaleSimple$breaks()}}
\item \href{#method-breaks_minor}{\code{ScaleSimple$breaks_minor()}}
\item \href{#method-labels}{\code{ScaleSimple$labels()}}
\item \href{#method-limits}{\code{ScaleSimple$limits()}}
\item \href{#method-trained_range}{\code{ScaleSimple$trained_range()}}
\item \href{#method-set_breaks}{\code{ScaleSimple$set_breaks()}}
\item \href{#method-set_breaks_minor}{\code{ScaleSimple$set_breaks_minor()}}
\item \href{#method-set_labels}{\code{ScaleSimple$set_labels()}}
\item \href{#method-set_limits}{\code{ScaleSimple$set_limits()}}
\item \href{#method-set_oob}{\code{ScaleSimple$set_oob()}}
\item \href{#method-set_range}{\code{ScaleSimple$set_range()}}
\item \href{#method-set_rescaler}{\code{ScaleSimple$set_rescaler()}}
\item \href{#method-set_trans}{\code{ScaleSimple$set_trans()}}
\item \href{#method-set_na_value}{\code{ScaleSimple$set_na_value()}}
\item \href{#method-set_limits_empty}{\code{ScaleSimple$set_limits_empty()}}
\item \href{#method-clone}{\code{ScaleSimple$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="map">}\href{../../ggr6/html/Scale.html#method-map}{\code{ggr6::Scale$map()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="name">}\href{../../ggr6/html/Scale.html#method-name}{\code{ggr6::Scale$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="set_guide">}\href{../../ggr6/html/Scale.html#method-set_guide}{\code{ggr6::Scale$set_guide()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ggr6" data-topic="Scale" data-id="within_limits">}\href{../../ggr6/html/Scale.html#method-within_limits}{\code{ggr6::Scale$within_limits()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$new(aesthetics = character(0))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_empty"></a>}}
\subsection{Method \code{is_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$transform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform"></a>}}
\subsection{Method \code{untransform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$untransform(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform_tbl"></a>}}
\subsection{Method \code{transform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$transform_tbl(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-untransform_tbl"></a>}}
\subsection{Method \code{untransform_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$untransform_tbl(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$train(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_tbl"></a>}}
\subsection{Method \code{train_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$train_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_tbl"></a>}}
\subsection{Method \code{map_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$map_tbl(data_trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks"></a>}}
\subsection{Method \code{breaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$breaks()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-breaks_minor"></a>}}
\subsection{Method \code{breaks_minor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$breaks_minor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-labels"></a>}}
\subsection{Method \code{labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-limits"></a>}}
\subsection{Method \code{limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$limits()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trained_range"></a>}}
\subsection{Method \code{trained_range()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$trained_range()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_breaks"></a>}}
\subsection{Method \code{set_breaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_breaks(breaks)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_breaks_minor"></a>}}
\subsection{Method \code{set_breaks_minor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_breaks_minor(breaks_minor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_labels"></a>}}
\subsection{Method \code{set_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_labels(labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_limits"></a>}}
\subsection{Method \code{set_limits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_limits(limits)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_oob"></a>}}
\subsection{Method \code{set_oob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_oob(oob)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_range"></a>}}
\subsection{Method \code{set_range()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_range(range)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rescaler"></a>}}
\subsection{Method \code{set_rescaler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_rescaler(rescaler)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_trans"></a>}}
\subsection{Method \code{set_trans()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_trans(trans)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_na_value"></a>}}
\subsection{Method \code{set_na_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_na_value(na_value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_limits_empty"></a>}}
\subsection{Method \code{set_limits_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$set_limits_empty(limits_empty)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScaleSimple$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
