% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-simple.R
\docType{data}
\name{ScaleSimple}
\alias{ScaleSimple}
\title{Default implementation of Scale objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ScaleSimple
}
\description{
Default implementation of Scale objects
}
\keyword{datasets}
