% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\docType{data}
\name{Stat}
\alias{Stat}
\alias{StatIdentity}
\title{Transformations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Stat

StatIdentity
}
\description{
Transformations
}
\keyword{datasets}
